requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
black==19.10b0
click==7.1.1
Django==3.0.5
djangorestframework==3.9.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
pylint==2.4.4
pytz==2019.3
regex==2020.4.4
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
wrapt==1.11.2
